pyproject.toml
[project]
name = "home-assistant-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "bitsandbytes>=0.45.5",
    "fastapi>=0.104.0",
    "paho-mqtt>=2.0.0",
    "qwen-omni-utils>=0.0.4",
    "qwen-vl-utils>=0.0.11",
    "requests>=2.32.3",
    "torchvision>=0.22.0",
    "transformers>=4.52.3",
    "uvicorn>=0.23.2",
    "vllm>=0.2.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers", rev = "v4.51.3-Qwen2.5-Omni-preview" }
